bot/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
aiogram==2.25.1
python-dotenv==1.0.1
pydantic==2.9.2